pyproject.toml
[project]
name = "screentranslate"
version = "0.1.0"
description = "Real-time screen region OCR and translation tool for games"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "mss>=10.1.0",
    "Pillow>=12.0.0",
    "pytesseract>=0.3.13",
    "deep-translator>=1.11.4",
    "pynput>=1.8.1",
    "gTTS>=2.5.4",
    "playsound3>=3.3.0",
    "pygame-ce>=2.5.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.1",
]

[project.scripts]
screentranslate = "screentranslate.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/screentranslate"]
